Add task to update PE Master group rules
This commit introduces a new private task to update
the AND conditional for the pe_compiler auth role in
the PE Master node group, changing it to regex match
for any *_compiler role. The task ensures that the
group rules are simplified and display more correctly
on the PE console.